refactor(foundation): move SemanticDriftResult + SemanticChangeType (A19)
Move the two semantic-drift data types from
`mockforge-core::ai_contract_diff::semantic_analyzer` to
`mockforge-foundation::contract_diff_types`:

- `SemanticDriftResult` (pure data: confidence/score, change type, LLM
  analysis JSON, before/after state, semantic mismatches)
- `SemanticChangeType` (enum: DescriptionChange, EnumNarrowing,
  NullableChange, ErrorCodeRemoved, SemanticConstraintChange,
  MeaningShift, SoftBreakingChange)

`SemanticAnalyzer` impl stays in core because it depends on
`OpenApiSpec` and `LlmClient`. Core re-exports the data types from
foundation for backwards-compat.

mockforge-http/handlers/semantic_drift.rs migrated to import both
types from foundation directly.

Drops another 11 deprecation warnings (298 → 287). All 574 http +
1818 core + 42 foundation tests pass; workspace clippy clean.
