+++ crates/mockforge-http/src/handlers/semantic_drift.rs
@@ -25,8 +25,8 @@ use crate::database::Database;
 fn map_row_to_semantic_incident(
     row: &sqlx::postgres::PgRow,
 ) -> Result<SemanticIncident, sqlx::Error> {
-    use mockforge_core::ai_contract_diff::semantic_analyzer::SemanticChangeType;
     use mockforge_core::incidents::types::{IncidentSeverity, IncidentStatus};
+    use mockforge_foundation::contract_diff_types::SemanticChangeType;
     use sqlx::Row;
 
     let id: uuid::Uuid = row.try_get("id")?;
